Updated the way the timer notifies the user when start and stop buttons are selected.
The user will now see that real world date/time the timer was started on.  When the user stops the timer they will see the duration of time passed since the last start added to the timer history.

Reversed the order of the history when saving a timer history to the database.